fix: modify getBaseQuery logic for getPagedBoards logic
In the SearchBoardRepositoryImpl class, we encountered an issue where heartCount and replyCount were returning the same values. This was due to the duplication of results in the join query. To resolve this, we separated the calculation of heartCount and replyCount into separate subqueries. Each subquery calculates the count of Heart and Reply for each Board independently.